+++ src/components/Results.vue
@@ -1,17 +1,12 @@
-<style>
-</style>
-
-<template>
-    <div v-if="hasResults">
-        <slot v-for="(result, index) in results" :result="result" :index="index">
-            {{ result }}
-        </slot>
-    </div>
-</template>
-
 <script>
     export default {
         inject: ['searchStore'],
+        props: {
+            tag: {
+                type: String,
+                default: 'div',
+            },
+        },
         data: function () {
             return {}
         },
@@ -19,9 +14,18 @@
             results() {
                 return this.searchStore.results
             },
-            hasResults() {
+            shouldRender() {
                 return this.results.length > 0
             },
+        },
+        render(h) {
+            if (!this.shouldRender) {
+                return ''
+            }
+
+            let slots = this.results.map(result => this.$scopedSlots.default({result}))
+
+            return h(this.tag, slots)
         },
     }
 </script>